requirements.txt
pandas
numpy
scikit-learn
xgboost
prophet
statsmodels
streamlit
openai
plotly
kaleido
cmdstanpy
azure-identity
azure-ai-ml
python-dotenv